reorganizes repo; reformats code
