+++ .github/workflows/lint.yml
@@ -17,7 +17,7 @@ jobs:
       - name: Install dependencies
         run: |
           python -m pip install --upgrade pip
-          python -m pip install black~=22.0 pylint
+          python -m pip install black~=22.0 pylint numpy
       - name: Analysing the code with pylint
         run: pylint --rcfile=.github/workflows/.pylintrc $(git ls-files '*.py')
       - name: Analysing the code with black
